+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install hardis-group/sfdx-hardis@beta`
 
-- check in manifest folder for package.xml
+- #24: Change the way of listing installed packages
+- #27: Check in manifest folder for package.xml
 
 ## [1.1.3] 2021-02-17